// src/app/core/models/userTest.model.ts
export class userTest {
    userId: string = '';
    username: string = 'Benade2024';
    password: string = '246750';
    completName: string = 'Benade Mabondzo';
    birthday: string = '03/09/1990';
    numeroRue: number = 10;
    rueName: string = 'Masseke';
    quartier: string = 'Fond Tié Tié';
    ville: string = 'Pointe-Noire';
    pays: string = 'Congo';
    ecoleId: string = 'b0f15d85-26f5-4002-ab15-a5c2178d17ce';
    profileId: string = 'admin';
}